+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Changed
+- All temporary status effects are removed from combatants at conclusion of combat.
+
 ## [0.3.34] - 2023-09-13
 ### Changed
 - Tweaked language ui.